Split TMap benchmark files
